+++ R/get_genome_ref.R
@@ -28,14 +28,16 @@
 #'
 #' @export
 #' @importFrom utils download.file unzip
+#' @importFrom tools R_user_dir
 #' 
 #' @examples
 #' \dontrun{
 #' genome_ref_path <- MAGMA.Celltyping::get_genome_ref()
 #' }
 get_genome_ref <- function(genome_ref_path = NULL,
-                           storage_dir = tools::R_user_dir("MAGMA.Celltyping",
-                                                           which="cache"),
+                           storage_dir = tools::R_user_dir(
+                               "MAGMA.Celltyping",
+                               which="cache"),
                            method = c("magma","piggback"),
                            population = c("eur", "afr", "amr", "eas", "sas"),
                            timeout = 60 * 5,